+++ src/runtime/bootstrap-custom-element.ts
@@ -67,13 +67,17 @@ export const proxyCustomElement = (Cstr: any, compactMeta: d.ComponentRuntimeMet
 
   const originalConnectedCallback = Cstr.prototype.connectedCallback;
   const originalDisconnectedCallback = Cstr.prototype.disconnectedCallback;
+  let hasHostListenerAttached = false;
   Object.assign(Cstr.prototype, {
     __registerHost() {
       registerHost(this, cmpMeta);
     },
     connectedCallback() {
-      const hostRef = getHostRef(this);
-      addHostEventListeners(this, hostRef, cmpMeta.$listeners$, false);
+      if (!hasHostListenerAttached) {
+        const hostRef = getHostRef(this);
+        addHostEventListeners(this, hostRef, cmpMeta.$listeners$, false);
+        hasHostListenerAttached = true;
+      }
 
       connectedCallback(this);
       if (BUILD.connectedCallback && originalConnectedCallback) {