+++ test/wdio/setup.ts
@@ -13,12 +13,14 @@ const testRequiresManualSetup =
   window.__wdioSpec__.includes('custom-elements-output-tag-class-different') ||
   window.__wdioSpec__.includes('custom-elements-delegates-focus') ||
   window.__wdioSpec__.includes('custom-elements-output') ||
+  window.__wdioSpec__.includes('no-external-runtime') ||
   window.__wdioSpec__.includes('global-script') ||
   window.__wdioSpec__.endsWith('custom-tag-name.test.tsx') ||
-  window.__wdioSpec__.endsWith('page-list.test.ts');
+  window.__wdioSpec__.endsWith('page-list.test.ts') ||
+  window.__wdioSpec__.endsWith('event-re-register.test.tsx');
 
 /**
- * setup all components defined in tests except for those where we want ot manually setup
+ * setup all components defined in tests except for those where we want to manually setup
  * the components in the test
  */
 if (!testRequiresManualSetup) {